pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=11.3.0",
    "gunicorn>=23.0.0",
    "django-cors-headers>=4.9.0",
    "openai>=1.61.0",
]
